backend/requirements.txt
# ============================================================================
# AI Business Portfolio Manager - Backend Dependencies
# ============================================================================
# Python 3.11+
# Last Updated: 2026-01-13

# -----------------------------------------------------------------------------
# Core Framework
# -----------------------------------------------------------------------------
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9

# -----------------------------------------------------------------------------
# Database & ORM
# -----------------------------------------------------------------------------
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# -----------------------------------------------------------------------------
# Redis & Caching
# -----------------------------------------------------------------------------
redis==5.0.1
hiredis==2.3.2

# -----------------------------------------------------------------------------
# Vector Database
# -----------------------------------------------------------------------------
qdrant-client==1.7.3

# -----------------------------------------------------------------------------
# AI & LLM
# -----------------------------------------------------------------------------
openai==1.12.0
anthropic==0.18.1
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.20
crewai==0.28.0
langgraph==0.0.40

# -----------------------------------------------------------------------------
# Task Queue & Background Jobs
# -----------------------------------------------------------------------------
celery==5.3.6
celery[redis]==5.3.6
flower==2.0.1
kombu==5.3.5

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
pandas==2.2.0
numpy==1.26.4
pyarrow==15.0.0

# -----------------------------------------------------------------------------
# Web Scraping & HTTP
# -----------------------------------------------------------------------------
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.2
praw==7.7.1  # Reddit API
vk-api==11.9.9  # VK API
python-telegram-bot==20.7  # Telegram

# -----------------------------------------------------------------------------
# GitHub Integration (Optional)
# -----------------------------------------------------------------------------
PyGithub==2.1.1

# -----------------------------------------------------------------------------
# Security & Authentication
# -----------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# -----------------------------------------------------------------------------
# Validation & Email
# -----------------------------------------------------------------------------
email-validator==2.1.0.post1

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dotenv==1.0.1
python-dateutil==2.8.2
pytz==2024.1
structlog==24.1.0
tenacity==8.2.3  # Retry logic
cachetools==5.3.2  # Caching utilities
pybloom-live==4.0.0  # Bloom filter

# -----------------------------------------------------------------------------
# Monitoring & Logging
# -----------------------------------------------------------------------------
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.40.0

# -----------------------------------------------------------------------------
# Rate Limiting
# -----------------------------------------------------------------------------
slowapi==0.1.9

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.6.0

# -----------------------------------------------------------------------------
# Development Tools
# -----------------------------------------------------------------------------
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
pre-commit==3.6.0

# -----------------------------------------------------------------------------
# Type Stubs
# -----------------------------------------------------------------------------
types-redis==4.6.0.20240106
types-pytz==2024.1.0.20240203
sqlalchemy[mypy]==2.0.27
